+++ src/NineLives/Views/ExposureView.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace Blackcat.NineLives.Views;
+
+public partial class ExposureView : UserControl
+{
+    public ExposureView()
+    {
+        InitializeComponent();
+    }
+}